--- Mixtape/__init__.py
@@ -1,1 +1,0 @@
-from param_sweep import param_sweep
