+++ Mixtape/utils.py
@@ -30,6 +30,10 @@ from sklearn.base import TransformerMixin
 from .base import BaseEstimator
 import sklearn.pipeline
 
+from sklearn import clone
+from sklearn.grid_search import ParameterGrid
+from sklearn.externals.joblib import Parallel, delayed
+
 #-----------------------------------------------------------------------------
 # Code
 #-----------------------------------------------------------------------------
@@ -206,3 +210,53 @@ class Subsampler(BaseEstimator, TransformerMixin):
             return [X[k::self._lag_time] for k in range(self._lag_time) for X in X_all]
         else:
             return [X[::self._lag_time] for X in X_all]
+
+
+def _fit_helper(args):
+    """
+    helper for fitting many models on some data
+    """
+    model, sequences = args
+    model.fit(sequences)
+
+    return model
+
+
+def param_sweep(model, sequences, param_grid, n_jobs=1):
+    """
+    Helper to fit many models to the same data
+
+    Parameters
+    ----------
+    model : mixtape.BaseEstimator
+        An *instance* of an estimator to be used
+        to fit data.
+    sequences : list of array-like
+        List of sequences, or a single sequence. Each 
+        sequence should be a 1D iterable of state
+        labels. Labels can be integers, strings, or
+        other orderable objects.
+    param_grid : dict or sklearn.grid_search.ParameterGrid
+        Parameter grid to specify models to fit. See
+        sklearn.grid_search.ParameterGrid for an explanation
+    n_jobs : int, optional
+        Number of jobs to run in parallel using joblib.Parallel
+
+    Returns
+    -------
+    models : list
+        List of models fit to the data according to
+        param_grid
+    """
+
+    if isinstance(param_grid, dict):
+        param_grid = ParameterGrid(param_grid)
+    elif not isinstance(param_grid, ParameterGrid):
+        raise ValueError("param_grid must be a dict or ParamaterGrid instance")
+
+    # iterable with (model, sequence) as items
+    iter_args = ((clone(model).set_params(**params), sequences) for params in param_grid)
+
+    models = Parallel(n_jobs=n_jobs)(delayed(_fit_helper)(args) for args in iter_args)
+
+    return models
